Unified Properties for Student & Grading Edition
Resolves #196
