+++ bundles/edu.kit.kastel.sdq.eclipse.grading.view/plugin.xml
@@ -14,22 +14,6 @@
             restorable="true">
       </view>
    </extension>
-   
-   <extension
-         point="org.eclipse.ui.preferencePages">
-      <page
-            class="edu.kit.kastel.eclipse.grading.view.preferences.ArtemisGradingPreferencesPage"
-            id="edu.kit.kastel.eclipse.grading.view.preferences.SamplePreferencePage"
-            name="Artemis Grading">
-      </page>
-   </extension>
-   <extension
-         name="Artemis Grading"
-         point="org.eclipse.core.runtime.preferences">
-      <initializer
-            class="edu.kit.kastel.eclipse.grading.view.preferences.PreferenceInitializer">
-      </initializer>
-   </extension>
 <extension
          point="org.eclipse.ui.perspectives">
          <perspective
